feat: add isInstanceof()
deprecates `isConstructor()`

`isType()` not accepting constructor anymore (breaking).

Updated project ot use changeset for better control of releases.
