+++ .github/mergify.yml
@@ -2,7 +2,7 @@ pull_request_rules:
   - name: automatic merge for Dependabot pull requests
     conditions:
       - author~=^dependabot(|-preview)\[bot\]$
-      - title~=^Bump [^\s]+ from ([\d]+)\..+ to \1\.
+      - title~=bump [^\s]+ from ([\d]+)\..+ to \1\.
     actions:
       merge:
         method: squash
