+++ .prettierrc.cjs
@@ -1,0 +1,11 @@
+module.exports = {
+  editorconfig: true,
+  printWidth: 110,
+  endOfLine: 'auto',
+  semi: false,
+  singleQuote: true,
+
+  // For ES5, trailing commas cannot be used in function parameters; it is counterintuitive
+  // to use them for arrays only
+  trailingComma: 'none'
+};
