+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "type-plus",
-  "version": "0.0.0-development",
+  "version": "4.18.1",
   "description": "Provides additional types for TypeScript.",
   "homepage": "https://github.com/unional/type-plus",
   "bugs": {
@@ -38,38 +38,38 @@
     "build:tslib": "tsc -p tsconfig.tslib.json && ncp package.cjs.json tslib/package.json",
     "build:doc": "typedoc",
     "clean": "rimraf cjs coverage esm lib libm tslib",
-    "coverage": "jest --coverage",
+    "coverage": "cross-env NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 jest --coverage",
+    "cs": "changeset",
     "depcheck": "depcheck",
     "lint": "eslint --ext=ts,js .",
     "nuke": "pnpm clean && rimraf node_modules",
     "postinstall": "husky install",
     "prepack": "pinst --disable",
     "postpack": "pinst --enable",
-    "release": "semantic-release",
+    "release": "changeset publish",
     "size": "size-limit",
-    "test": "jest",
-    "test:watch": "jest --watch",
+    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 jest",
+    "test:watch": "cross-env NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 jest --watch",
     "verify": "npm-run-all clean -p build depcheck lint coverage -p size",
+    "version": "changeset version",
     "watch": "jest --watch"
   },
   "dependencies": {
-    "tersify": "^3.10.2",
-    "unpartial": "^1.0.3"
+    "tersify": "^3.11.1",
+    "unpartial": "^1.0.4"
   },
   "devDependencies": {
+    "@changesets/cli": "^2.25.2",
     "@commitlint/cli": "^17.0.2",
     "@commitlint/config-conventional": "^17.0.2",
     "@repobuddy/jest": "^2.1.0",
-    "@semantic-release/changelog": "^6.0.1",
-    "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^8.0.5",
-    "@semantic-release/npm": "^9.0.1",
     "@size-limit/preset-small-lib": "^8.0.0",
     "@types/jest": "^29.0.0",
     "@types/node": "^14",
     "@typescript-eslint/eslint-plugin": "^5.28.0",
     "@unional/fixture": "^3.0.0",
     "assertron": "^11.0.1",
+    "cross-env": "^7.0.3",
     "depcheck": "^1.4.3",
     "eslint": "^8.17.0",
     "eslint-plugin-harmony": "^7.1.0",
@@ -83,7 +83,6 @@
     "pinst": "^3.0.0",
     "rimraf": "^3.0.2",
     "satisfier": "^5.2.2",
-    "semantic-release": "^19.0.3",
     "size-limit": "^8.0.0",
     "ts-jest": "^29.0.0",
     "tslib": "^2.4.0",
