+++ ts/class/isConstructor.spec.ts
@@ -1,4 +1,4 @@
-import { AnyConstructor, assertType, isConstructor } from '../index.js'
+import { AnyConstructor, assertType, isConstructor, isInstanceof } from '../index.js'
 
 test('check constructor', () => {
   const x: unknown = Error
@@ -32,4 +32,22 @@ test('constructor throwing undefined', () => {
 test('class with params returns true', () => {
   class Boo { constructor(public a: number) { } }
   expect(isConstructor(Boo)).toBe(true)
+})
+
+describe('isInstanceof()', () => {
+  test('Class as validator', () => {
+    const s: unknown = new Error()
+    if (isInstanceof(s, Error))
+      assertType<Error>(s)
+  })
+  test('Class as validator fails', () => {
+    class Foo { a() { } }
+    const s: Foo | number = 1
+    if (isInstanceof(s, Foo)) {
+      assertType<Foo>(s)
+    }
+    else {
+      assertType<number>(s)
+    }
+  })
 })
