+++ tsconfig.base.json
@@ -7,6 +7,7 @@
       "ES2015",
       "ES2021.Promise"
     ],
+    "isolatedModules": true,
     "newLine": "LF",
     "noUnusedLocals": true,
     "noUnusedParameters": false,
